+++ vkt/gpu/vulkan/vulkan_surface_android.cpp
@@ -22,7 +22,7 @@ void VulkanSurface::createSurfaceKHR()
                                                             &m_surface);
     if (result != VK_SUCCESS)
     {
-        throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", result));
+        throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", static_cast<int32_t>(result)));
     }
 }
