+++ vkt/gpu/vulkan/vulkan_surface_macos.mm
@@ -41,7 +41,7 @@ void VulkanSurface::createSurfaceKHR()
 
         if (result != VK_SUCCESS)
         {
-            throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", result));
+            throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", static_cast<int32_t>(result)));
         }
     }
 }
@@ -75,7 +75,7 @@ void VulkanSurface::createSurfaceKHR()
 
         if (result != VK_SUCCESS)
         {
-            throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", result));
+            throw std::runtime_error(fmt::format("Failed to create VkSurfaceKHR.: {}", static_cast<int32_t>(result)));
         }
     }
 }
